change integer to bigint for braintree IDs

+++ database/migrations/2021_04_14_151911_add-user-additional-fields.php
@@ -15,7 +15,7 @@ class AddUserAdditionalFields extends Migration
     {
         Schema::table('users', function(Blueprint $table) {
             $table->string('customer_id')->nullable();
-            $table->integer('braintree_customer_id')->nullable();
+            $table->bigInteger('braintree_customer_id')->nullable();
         });
     }
